Refactor `generate_transaction_status_message`
