+++ ledger_bot/message_generators/__init__.py
@@ -1,0 +1,6 @@
+"""Functions for generating message content posted by ledger-bot."""
+from . import generate_transaction_status_message
+
+generate_transaction_status_message = (
+    generate_transaction_status_message.generate_transaction_status_message
+)
