+++ ledger_bot/process_transactions/mark_transaction_paid.py
@@ -5,7 +5,7 @@ import logging
 
 import discord
 
-from ledger_bot.message_generator import generate_transaction_status_message
+from ledger_bot.message_generators import generate_transaction_status_message
 from ledger_bot.models import Transaction
 from ledger_bot.storage import AirtableStorage
